criptografar senha , refatorar codigo

+++ ClienteCrud/UsuarioRepositorioComBanco.cs
@@ -11,19 +11,9 @@ using System.Threading.Tasks;
 
 namespace ClienteCrud
 {
-    public class UsuarioRepositorioComBanco : IDisposable
+    public class UsuarioRepositorioComBanco
     {
         private static SqlConnection sqlConexao;
-        private static string sqlConexaoString;
-
-        public UsuarioRepositorioComBanco()
-        { }
-        private static string bancoConexaoString()
-        {
-            sqlConexaoString = ConfigurationManager.ConnectionStrings
-                ["conexaoSql"].ConnectionString;
-            return sqlConexaoString;
-        }
         private static SqlConnection BancoConexao()
         {
             sqlConexao = new SqlConnection(ConfigurationManager.ConnectionStrings
@@ -33,57 +23,56 @@ namespace ClienteCrud
         }
 
 
-        public DataTable ObterTodos()
+        public List<Usuario> ObterTodos()
         {
-            SqlDataAdapter comandoBancoDeDados = null;
             DataTable BancoDeDados = new DataTable();
             using(var conexao = BancoConexao())
             {
                 using (var cmd = conexao.CreateCommand())
                 {
                     cmd.CommandText = "SELECT * FROM USUARIO";
-                    comandoBancoDeDados = new SqlDataAdapter(cmd.CommandText, conexao);
+                    var comandoBancoDeDados = new SqlDataAdapter(cmd.CommandText, conexao);
                     comandoBancoDeDados.Fill(BancoDeDados);
-                    return BancoDeDados;
                 }
-            }
-
-        }
-        public List<Usuario> ConverterDataTableParaUsuario()
-        {
-            var repositorioComBanco = new UsuarioRepositorioComBanco();
-            var conventerDataTableParaUsuario = new ConversorDataTableParaUsuario(); 
-            return conventerDataTableParaUsuario.ConverterParaLista<Usuario>(repositorioComBanco.ObterTodos()); 
+            } 
+            return ConversorDataTableParaUsuario.ConverterParaLista<Usuario>(BancoDeDados);
         }
 
         public void AdicionarUsuario(Usuario usuario)
         {
-            using (var cmd = BancoConexao().CreateCommand())
-            {
-                cmd.CommandText = "INSERT INTO USUARIO (NOME, SENHA, EMAIL, DATACRIACAO, DATANASCIMENTO) VALUES (@NOME, @SENHA, @EMAIL, @DATACRIACAO, @DATANASCIMENTO)";
-                cmd.Parameters.AddWithValue("@NOME", usuario.Nome);
-                cmd.Parameters.AddWithValue("@SENHA", usuario.Senha);
-                cmd.Parameters.AddWithValue("@EMAIL", usuario.Email);
-                cmd.Parameters.AddWithValue("@DATACRIACAO", usuario.DataCriacao);
-                if (usuario.DataNascimento == null)
+            using (var conexao = BancoConexao()) 
+            { 
+                using(var cmd = conexao.CreateCommand())
                 {
-                    cmd.Parameters.AddWithValue("@DATANASCIMENTO", DBNull.Value);
-                }
-                else
-                {
-                    cmd.Parameters.AddWithValue("@DATANASCIMENTO", usuario.DataNascimento);
+                     cmd.CommandText = "INSERT INTO USUARIO (NOME, SENHA, EMAIL, DATACRIACAO, DATANASCIMENTO) VALUES (@NOME, @SENHA, @EMAIL, @DATACRIACAO, @DATANASCIMENTO)";
+                     cmd.Parameters.AddWithValue("@NOME", usuario.Nome);
+                     cmd.Parameters.AddWithValue("@SENHA",  CriptografarSenha.Criptografar(usuario.Senha));
+                     cmd.Parameters.AddWithValue("@EMAIL", usuario.Email);
+                     cmd.Parameters.AddWithValue("@DATACRIACAO", usuario.DataCriacao);
+                     if (usuario.DataNascimento == null)
+                     {
+                         cmd.Parameters.AddWithValue("@DATANASCIMENTO", DBNull.Value);
+                     }
+                     else
+                     {
+                         cmd.Parameters.AddWithValue("@DATANASCIMENTO", usuario.DataNascimento);
 
+                     }
+                     cmd.ExecuteNonQuery();
                 }
-                cmd.ExecuteNonQuery();
-                sqlConexao.Close();
             }
         }
 
         public void EditarUsuario(Usuario usuario)
         {
-            using (var cmd = BancoConexao().CreateCommand())
+            if (usuario == null)
             {
-                if(usuario != null)
+                throw new ArgumentNullException("Usuario não pode ser nulo");
+            }
+
+            using (var conexao = BancoConexao())
+            {
+                using(var cmd = conexao.CreateCommand())
                 {
                     cmd.CommandText = "UPDATE USUARIO SET NOME=@NOME, SENHA=@SENHA, EMAIL=@EMAIL, DATACRIACAO=@DATACRIACAO, DATANASCIMENTO=@DATANASCIMENTO WHERE ID=@ID";
                     cmd.Parameters.AddWithValue("ID", usuario.Id);
@@ -101,28 +90,28 @@ namespace ClienteCrud
 
                     }
                     cmd.ExecuteNonQuery();
-                    sqlConexao.Close();
                 }
             }
         }
 
         public void RemoverUsuario(int id)
         {
-            var usuario = new Usuario();
-            using (var cmd = BancoConexao().CreateCommand())
+            if (id == decimal.Zero)
             {
-                cmd.CommandText = "DELETE FROM USUARIO WHERE ID=@ID";
-                cmd.Parameters.AddWithValue("@ID", id);
-                cmd.ExecuteNonQuery();
-                sqlConexao.Close();
+                throw new ArgumentNullException("Id não pode ser zero");
+            }
+
+            using (var conexao = BancoConexao())
+            {
+
+                using(var cmd = conexao.CreateCommand())
+                {
+                    cmd.CommandText = "DELETE FROM USUARIO WHERE ID=@ID";
+                    cmd.Parameters.AddWithValue("@ID", id);
+                    cmd.ExecuteNonQuery();
+                }
             }
         }
-
-        public void Dispose()
-        {
-            GC.SuppressFinalize(this);
-        }
-
         
     }
 }
